requirements.txt
asgiref==3.8.1
Django==5.1.1
sqlparse==0.5.1
python-dotenv~=1.0.1
mysqlclient==2.2.4
pillow==10.4.0
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
pytest==8.3.3
pytest-django==4.9.0
djoser==2.2.3



